requirements.txt
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.2.4/en_ner_bc5cdr_md-0.2.4.tar.gz
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.2.4/en_core_sci_md-0.2.4.tar.gz
opencv-python==4.5.1.48
nltk==3.4.5
torchtext==0.5.0
Pillow==7.0.0
Jinja2==2.11.1
imageio==2.6.1
scikit-image==0.16.2
scikit-learn==0.22.1
scispacy==0.2.4
pydicom==1.4.2
pysftp==0.2.9
pandas==1.1.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-WTF==0.14.3
Django==3.0.5
WTForms==2.2.1
Werkzeug==1.0.0
python-crfsuite==0.9.7
marisa-trie==0.7.5
repoze.lru==0.7
googletrans==3.1.0a0
